(maint) Use smaller unique identifier for clone hostname
The clone class attempts to set up a unique hostname for each clone by
appending the base certname to the clone name in an /etc/hostname file
for each clone overlay. This works so long as the combination of
title.certname does not exceed 64 characters. If the base node has a
long certname then you have two failure cases.

1. the truncated certname ends up not being unique across nodes, which
   is only a problem if you are using multiple clone army nodes
2. the truncated certname ends on a forbidden character like a hyphen,
   in which case it will be rejected by puppetserver.

To try and work around this, I'm substituting the ip of the node for the
certname.